+++ balances-tx-payment-extension/src/lib.rs
@@ -1,0 +1,6 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod balance_conversion;
+
+#[cfg(test)]
+mod tests;
